Update starting-utils.md
Duplicated note on connecting to Internet using Ethernet in the "Installing solo-utils" section. I don't like the duplication but this is the first time that we suggest connecting to both Solo and the host network (so it needs to be here). I think it also needs to be said below.